ft_lstsize: Changed return type to 'size_t'

+++ 42-cursus/libft/ft_lstsize_bonus.c
@@ -6,13 +6,13 @@
 /*   By: rtorrent <rtorrent@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/09 13:46:32 by rtorrent          #+#    #+#             */
-/*   Updated: 2023/05/12 16:23:06 by rtorrent         ###   ########.fr       */
+/*   Updated: 2023/06/18 14:11:30 by rtorrent         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-static int	g_size;
+static size_t	g_size;
 
 static void	f_aux(void *content)
 {
@@ -20,7 +20,7 @@ static void	f_aux(void *content)
 	g_size++;
 }
 
-int	ft_lstsize(t_list *lst)
+size_t	ft_lstsize(t_list *lst)
 {
 	g_size = 0;
 	ft_lstiter(lst, f_aux);
